Create LegendaryFarming
You’ve beaten all the content and the last thing left to accomplish is to own a legendary item. However, it’s a tedious process and requires quite a bit of farming. Anyway, you are not too pretentious – any legendary will do. The possible items are:
•	"Shadowmourne" – requires 250 Shards
•	"Valanyr" – requires 250 Fragments
•	"Dragonwrath" – requires 250 Motes
"Shards", "Fragments", and "Motes" are the key materials, all else is junk. 
You will be given lines of input in the format: 
"{quantity1} {material1} {quantity2} {material2} … {quantityN} {materialN}"
Keep track of the key materials - the first that reaches the 250 mark wins the race. At that point, print the corresponding legendary obtained. 
Then, print the remaining shards, fragments, motes, ordered by quantity in descending order, then by name in ascending order, each on a new line. Finally, print the collected junk items, in alphabetical order.