Add files via upload
On the first line, you will receive the quantity of food, which Merry buys for a month (30 days). On the second line, you will receive quantity hay, and on the third line – quantity cover, also for a month. On the fourth line, you will receive the guinea pig's weight.  All values are in kilograms.
 
Calculate whether the quantity of food, hay, and cover, will be enough for a month. The amount of food given to the guinea pig each day is 300 gr. Every second day Merry gives hay, which is 5% of the rest of the food. On every third day, Merry put cover, which is 1/3 of the guinea pig's weight. 
 
Puppy eats food every single day!
 
If food, hay, or cover runs out, stop the program.