Add files via upload
You will receive an initial list with groceries separated by "!".

After that you will be receiving 4 types of commands, until you receive "Go Shopping!"

Urgent {item} - add the item at the start of the list.  If the item already exists, skip this command.
Unnecessary {item} - remove the item with the given name, only if it exists in the list. Otherwise skip this command.
Correct {oldItem} {newItem} – if the item with the given old name exists, change its name with the new one. If it doesn't exist, skip this command.
Rearrange {item} - if the grocery exists in the list, remove it from its current position and add it at the end of the list.